Move MessageParser the clp namespace.
